chore: BaseProblem is inherited by sub problems
